requirements.txt
pillow==3.1.1
funcy==1.7
matplotlib==1.5.1
pandas==0.17.1
scipy==0.17.0
